fix: Prevent passing invalid values for next_offset
Arroyo's next_offset (used for seeking and committing on the consumer)
must always be equal to the offset of the last processed message + 1.
We should not allow any other value to be passed here by the user.
We should also update the Python version of the library to do the same.