feat: Poll duration for Kafka backend
If a poll duration is passed to the Kafka backend, it is now respected
rather than using the hardcoded 1 second value.

The default poll timeout for the Kafka backend is also updated to 100 milliseconds
instead of 1 second to match the default in the Python arroyo library.

This also changes the signature of the poll method to take a duration instead
of a float that needs to be converted to a duration each time we call poll.
Passing a duration seems less ambiguous that passing an f64 here.