ref: Remove lifetime annotations from ArroyoConsumer
This is an attempt to simplify the ArroyoConsumer lifetimes by
changing the LocalConsumer to have ownership of the LocalBroker.

Since this consumer/broker only used in place of the KafkaConsumer
for testing, I think this change is fairly harmless as it's not
required for the broker to live outside the consumer.