feat(errors) Redesign errors management step 1  (#23)
This start introducing thiserror (https://crates.io/crates/thiserror) as a way to rationalize the errors management, that now is useless as errors do not carry the cause and are used inconsistently.

This starts by defining new errors for the consumer trait and for the two implementations: local and kafka.

Add a new enum for all meaningful error. This is an api so errors tend to be fairly transparent.
Just makes the compilation work.