test: Add test for commit function (#35)
Adds an end to end test for committing offsets in the Kafka consumer. Also adds helper functions for creating and deleting Kafka topics. Needs a running Kafka on `localhost:9092` to work.

Now we run Kafka and Zookeeper in CI. Since docker is not available in macos, removed that test suite.