feat: on_assign callback is fired post rebalance (#51)
Now we fire the on_assign callback post-rebalance rather than pre-rebalance.
on_revoke still gets fired pre-rebalance. This is so on_assign function
will not error if the user tries to do something with the partition that
is not assigned yet.

Hopefully also fixes the flaky commit offsets test as consumer.tell() now
only returns the partition after it is assigned and not before.